adjusted public api service tests to changes in client folder structure
