+++ projects/services/public-api-service-typescript/.gitignore
@@ -1,2 +1,3 @@
 # Ignore built ts files
 dist/**/*
+src/orbs-api-interface.ts
