+++ projects/services/public-api-service-typescript/build.sh
@@ -6,5 +6,8 @@ yarn link orbs-core-library
 yarn install
 
 yarn run build
+a
+echo "Copy orbs-api-interface from contract test..."
+cp ../../../client/client-contract-test/src/orbs-api-interface.ts src
 
 yarn test
